requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
Pillow
gunicorn